+++ src/app/_components/card/QuizCard.tsx
@@ -3,11 +3,15 @@
 import React from "react";
 import { Card, CardBody } from "@nextui-org/react";
 import MyButton from "app/_components/MyButton";
+import AnswerOptions from "app/_components/quiz/AnswerOptions";
 
 interface QuizCardProps {
     leftStreak: number;
-    courseTitle: string;
+    subject: string;
     question: string;
+    options: string[];
+    onSelect: (selected: string) => void;
+    selected: string | null;
 }
 
 const handleAskAI = () => {
@@ -25,8 +29,11 @@ const handleSubmit = () => {
 
 const QuizCard: React.FC<QuizCardProps> = ({
                                                leftStreak,
-                                               courseTitle,
+                                               subject,
                                                question,
+                                               options,
+                                               onSelect,
+                                               selected,
                                            }) => {
     return (
         <Card className="w-full max-w-xl p-4 min-w-[800px]">
@@ -34,9 +41,11 @@ const QuizCard: React.FC<QuizCardProps> = ({
                 <div className="bg-lime-400 rounded-lg px-4 py-1 mb-4 w-64 h-[26px] text-white">
                     <p><b>{leftStreak}문제</b>만 더 풀면 스트릭을 쌓아요</p>
                 </div>
-                <p className="text-xs text-gray-400 mb-1">한입코딩 &amp;gt 코스 &amp;gt {courseTitle}</p>
+                <p className="text-xs text-gray-400 mb-1">한입코딩 > 코스 > {subject}</p>
                 <h2 className="text-lg font-semibold mb-4">{question}</h2>
-                <div className="flex justify-center mt-8 gap-4">
+                <AnswerOptions options={options} onSelect={onSelect} selected={selected} />
+                <div className="flex justify-start"></div>
+                <div className="flex justify-end mt-8 gap-4">
                     <MyButton className="bg-purple-500 hover:bg-purple-600" onClick={handleAskAI}>
                         AI에게 물어보기
                     </MyButton>
